chore: minor test updates and refactor based on pr feedback
